Update photoz_helper.py
This adds an assertion statement to help clarify what I think would be a confusing error to a user. More generally... I think this is fixed by Kostoya's change since the user will be exposed to the DF I create as a result of the PanStarrs query, which will make the query results directly explorable. When that query fails completely for all Hosts, then this will fail appropriately.